Data.Scoped

Description

A value that has local scope and global scope

Documentation

newtype Interval a

A closed interval

Constructors

Interval

Fields

getInterval :: (a, a)

_Interval :: forall a a. Iso (Interval a) (Interval a) (a, a) (a, a)

data Scoped a

A value scoped in local interval and global interval

Constructors

Scoped

Fields

_scoped :: a

_locally :: Interval a

_globally :: Interval a

scoped :: forall a. Lens' (Scoped a) a

locally :: forall a. Lens' (Scoped a) (Interval a)

globally :: forall a. Lens' (Scoped a) (Interval a)

rangeScope :: [a] -> Int -> Maybe (Scoped Int)

rangeOf :: Ix a => Scoped a -> [a]

scopedTo :: (Num a, Ord a, Show a) => a -> a -> Scoped a

adjustTo0 :: Num a => Scoped a -> Scoped a

localIx :: Num a => Getter (Scoped a) a

forward :: (Num a, Ix a) => Scoped a -> Scoped a

Forwarding stickey

back :: (Num a, Ix a) => Scoped a -> Scoped a

Backwarding stickey