SDLight.Util

Description

Utility module

Color

class RGBA c where

Represents RGB with alpha value

Minimal complete definition

rgba, rgb, opacity

Methods

rgba :: Int -> Int -> Int -> Int -> c

rgb :: Int -> Int -> Int -> c

opacity :: Int -> c -> c

newtype Color

Color datatype represented as four sequences of number in [0,255]

Constructors

Color (V4 Int)

getColor :: Color -> V4 Int

Destructor of Color

black :: Color

black = rgba 0 0 0 255

white :: Color

white = rgba 255 255 255 255

red :: Color

red = rgba 255 0 0 255

green :: Color

green = rgba 0 255 0 255

blue :: Color

blue = rgba 0 0 255 255

orange :: Color

orange = rgba 255 128 0 255

Rectangle

_position :: Lens' (Rectangle a) (Point V2 a)

Accessor the upper left corner of a rectangle

_size :: Lens' (Rectangle a) (V2 a)

Accessor the size of a rectangle

Lenses

functorial :: Functor f => Getter a b -> Getter (f a) (f b)

functorial lifting

monadic :: Monad m => Lens' a b -> Lens' (m a) (m b)

monadic lifting

(^%~) :: Lens' s a -> Getter a a -> s -> s infixr 4

s ^%~ f = s %~ (^. f)

(^%%~) :: Functor m => Lens' s a -> Getter a (m a) -> s -> m s infixr 4

s ^%%~ f = s %%~ (^. f)

(^%=) :: MonadState s m => Lens' s a -> Getter a a -> m () infixr 4

s ^%= f = s %= (^. f)

(^%%=) :: MonadState s m => Lens' s a -> Getter a (m a) -> m () infixr 4

s ^%%= f = use s >>= \s' -> s <~ s' ^. f

Extensible

hmerge :: (xs ⊆ ys, Wrapper h) => (h :* xs) -> (h :* ys) -> h :* ys

Merges two products when first one is included in second one

hmergeAssoc :: (IncludeAssoc ys xs, Wrapper h) => (h :* xs) -> (h :* ys) -> h :* ys

Merges two associated products when first one is included in second one