SDLight.Widgets.Animated

Description

Animation layer from sprite image file

Widget

wAnimated :: Given StyleSheet => WConfig "animated" -> GameM (Widget Op'Animated)

Animated widget

Animation goes from left in the first line to right in the first line, and goes in the second line, and so on.

Config Parameter

Required

[ "texture" >: SDL.Texture  -- A texture connected horizontally and/or vertically
, "pictureSize" >: V2 Int  -- Size of each frame
]

Optional

[]

Methods

op'render Render operator

op'run Tick an animation by one frame

Method

type Op'Animated = '[Op'Render, Op'Run]

Method of wAnimated

Orphan instances

Conf Symbol "animated"

Associated Types

type Required "animated" (k :: "animated") :: [Assoc Symbol *]

type Optional "animated" (k :: "animated") :: [Assoc Symbol *]

Methods

def :: Record Symbol (Optional "animated" k)