SDLight.Widgets.Core

Description

Core of Widget

Widget functions

override :: (Widget old -> Widget new) -> Widget old -> (forall br m. Union new br m ~> (br (Widget new) m `Sum` Union old br m)) -> Widget new

recursive extension

runSwitch :: Widget xs -> Getter (Widget xs) (FreezeT (Widget xs) Identity a) -> (Freeze (Widget xs) a -> r) -> r

combinator for running a widget with switching method like op'switch

runSwitchM :: (k ∈ xs, Monad m) => Widget xs -> k FreezeT m a -> (Freeze (Widget xs) a -> m r) -> m r

runSwitch with monad

op'isFreeze :: Widget xs -> Getter (Widget xs) (FreezeT (Widget xs) Identity a) -> Bool

Is a widget frozen?

Config

type WConfig k = WConfigR (Record (Required k)) (Record (Optional k))

WConfig k is a record of widget id, required parameter and optional parameter

class Conf k where

k has default value of optional parameters

Minimal complete definition

def

Associated Types

type Required k :: [Assoc Symbol *]

type Optional k :: [Assoc Symbol *]

Methods

def :: Record (Optional k)

default values

data ViewWConfig a b

Unpack a value of WConfig

Constructors

ViewWConfig WidgetId a b

giveWid :: KnownSymbol k => Proxy k -> WConfig k -> WConfig k

Give a widget-id using Proxy

getLocation :: Given StyleSheet => WidgetId -> V2 Int

Find the location of the widget corresponding to given widget-id

mkCfg :: WidgetId -> a -> b -> WConfigR a b

Build Config data

viewWConfig :: WConfigR a b -> ViewWConfig a b

Unpacking function

wconf :: KnownSymbol k => Proxy k -> WConfig k -> ViewWConfig (Record (Required k)) (Record (Optional k))

wconf p = viewWConfig . giveWid p

conf :: WidgetId -> Record (Required k) -> Record (Optional k) -> WConfig k

combinator for a widget with wix (widget-id)

  conf wix req opt
  = #wix = wix
  <: #required = req
  <: #optional @= opt
  <: emptyRecord

Common Operators

data Op'Render br m r where

Constructors

Op'Render :: Double -> Op'Render Value GameM ()

op'render :: (Given StyleSheet, KnownName xs, Op'Render ∈ xs) => Getter (Widget xs) (GameM ())

op'render = op'renderAlpha 1.0

op'renderAlpha :: (KnownName xs, Op'Render ∈ xs) => Double -> Getter (Widget xs) (GameM ())

data Op'Run br m r where

Constructors

Op'Run :: Op'Run Self GameM a

op'run :: Op'Run ∈ xs => Getter (Widget xs) (GameM (Widget xs))

op'run = _self Op'Run

data Op'Reset arg br m r where

Constructors

Op'Reset :: arg -> Op'Reset arg Self Identity a

op'reset :: Op'Reset arg ∈ xs => arg -> Getter (Widget xs) (Widget xs)

op'reset = _self' . Op'Reset

data Op'HandleEvent br m r where

Constructors

Op'HandleEvent :: Map Scancode Int -> Op'HandleEvent Self GameM a

op'handleEvent :: Op'HandleEvent ∈ xs => Map Scancode Int -> Getter (Widget xs) (GameM (Widget xs))

op'handleEvent = _self . Op'HandleEvent

data Op'Switch br m r where

Constructors

Op'Switch :: Op'Switch FreezeT Identity ()

op'switch :: Op'Switch ∈ xs => Getter (Widget xs) (FreezeT (Widget xs) Identity ())

op'switch = _Op Op'Switch

Re-exports

type family (k ++ (xs :: [k])) (ys :: [k]) :: [k] where ... infixr 5

Equations

(k ++ ([] k)) ys = ys

(k ++ ((:) k x xs)) ys = (:) k x ((++) k xs ys)

module SDLight.Widgets.Internal.Widget

module SDLight.Widgets.Internal.TH

module SDLight.Widgets.Internal.Named

module SDLight.Widgets.Internal.Freeze