SDLight.Widgets.Internal.Freeze

Description

Freeze datatype

Documentation

data Freeze w a

Freeze w a represents a datatype that keeps w normally or w is frozen with a value of a

Constructors

Freeze w a

Keep w

_Keep :: forall w a. Prism' (Freeze w a) w

_Freeze :: forall w a a. Prism (Freeze w a) (Freeze w a) (w, a) (w, a)

refreeze :: (w -> z) -> Freeze w a -> Freeze z b

Mapping function on first argument

unfreeze :: (w -> a -> r) -> (w -> r) -> Freeze w a -> r

Case-analysis on Freeze w a

isFreeze :: Freeze w a -> Bool

Is this frozen?

_Frozen :: Lens' (Freeze w a) w

Lens used for accessing its internal value w

newtype FreezeT w m a

Freeze value with monad

Constructors

FreezeT

Fields

runFreezeT :: m (Freeze w a)

freeze :: Monad m => Widget xs -> a -> FreezeT (Widget xs) m a

freezeM with Identity monad

freeze' :: Monad m => Widget xs -> FreezeT (Widget xs) m ()

freeze with no value

freezeM :: Monad m => m (Widget xs) -> a -> FreezeT (Widget xs) m a

Freeze given widget with a value

freezeM' :: Monad m => m (Widget xs) -> FreezeT (Widget xs) m ()

freezeM with no value

onFreeze :: (TransBifunctor FreezeT m, Monad m) => Lens' s (Widget xs) -> (Widget xs -> FreezeT (Widget xs) Identity a) -> s -> (a -> Widget xs -> m s) -> m s

Call a method op and if the result is frozen, pass its value to callback function cb. Otherwise update the widget object.

onFreeze' :: (TransBifunctor FreezeT m, Monad m) => Lens' s (Widget xs) -> (Widget xs -> FreezeT (Widget xs) Identity ()) -> s -> (Widget xs -> m s) -> m s

onFreeze with no value

onFinish :: Monad m => Lens' s (Widget xs) -> (Widget xs -> FreezeT (Widget xs) Identity a) -> s -> (a -> s -> m s) -> m s

onFinish lens op s cb = onFreeze lens op s (a w -> cb a (s & lens .~ w))

onFinishM :: (TransBifunctor FreezeT m, Monad m) => Lens' s (Widget xs) -> (Widget xs -> m (FreezeT (Widget xs) Identity a)) -> s -> (a -> s -> m s) -> m s

Monadic function of onFinish