SDLight.Widgets.Internal.Named

Description

Named Widget

Documentation

data Op'GetName br m r where

Constructors

Op'GetName :: Op'GetName Value Identity WidgetId

op'getName :: (∈) Op'GetName xs => Getter (Widget xs) WidgetId

type Named xs = Op'GetName ': xs

Named xs is a method list which contains Op'GetName

type NamedWidget xs = Widget (Named xs)

NamedWidget xs widget has its own name with a method list xs

wNamed :: WidgetId -> Widget xs -> Widget (Named xs)

named widget

Methods

op'getName

class KnownName xs where

Get its widget id from a widget Returns Nothing if xs does not have Op'GetName method

Minimal complete definition

symbolName

Methods

symbolName :: Widget xs -> Maybe WidgetId