SDLight.Widgets.Internal.Widget

Description

Definition of Widget

Documentation

type (~>) f g = forall x. f x -> g x

Polymorphic function type

newtype Op br m r op

Operator type

Constructors

Op

Fields

runOp :: op br m r

newtype Union ops br m v

Union ops br m v represents operator list of ops with each operator has type of Op br m v

Constructors

Union

Fields

getUnion :: Op br m v :| ops

newtype Widget ops

Widget Type

Constructors

Widget

Fields

runWidget :: forall br m. (Functor m, TransBifunctor br m) => Union ops br m ~> br (Widget ops) m

class TransBifunctor f m where

f is like a bifunctor with base monad m

Minimal complete definition

bimapT

Methods

bimapT :: (a -> b) -> (c -> d) -> f a m c -> f b m d

firstT :: (a -> b) -> f a m c -> f b m c

secondT :: (c -> d) -> f a m c -> f a m d

call :: (k ∈ xs, TransBifunctor br m, Functor m) => Widget xs -> k br m ~> br (Widget xs) m

Call a method of a widget

_Op :: (k ∈ xs, TransBifunctor br m, Functor m) => k br m a -> Getter (Widget xs) (br (Widget xs) m a)

Turn a operator into Getter

(@>) :: (t br m ~> r) -> (Union ts br m ~> r) -> Union (t ': ts) br m ~> r infixr 2

Cons operator of Union type

emptyUnion :: Union '[] br m v -> a

Exhausts empty union

newtype Self w m a

Represents that an operator returns w, usually widget itself

Constructors

Self

Fields

runSelf :: m w

newtype Value w m a

Represents that an operator returns a, usually a value

Constructors

Value

Fields

getValue :: m a

class NodeW br where

br can continue with the current widget state

Minimal complete definition

continue, continueM

Methods

continue :: Monad m => Widget xs -> br (Widget xs) m a

continueM :: Functor m => m (Widget xs) -> br (Widget xs) m a

_self :: (k ∈ xs, TransBifunctor br m, Functor m) => k br m () -> Getter (Widget xs) (m (Widget xs))

class LeafW br where

br can finish with given value

Minimal complete definition

finish, finishM

Methods

finish :: Monad m => model -> br (Widget xs) m model

finishM :: Functor m => m model -> br (Widget xs) m model

_value :: (k ∈ xs, Functor m) => k br m a -> Getter (Widget xs) (m a)

_self' :: (k ∈ xs, TransBifunctor br Identity, NodeW br) => k br Identity () -> Getter (Widget xs) (Widget xs)

_self with Identity monad

_value' :: (k ∈ xs, LeafW br) => k br Identity a -> Getter (Widget xs) a

_value' with Identity monad