SDLight.Widgets.ScriptEngine

Description

ScriptEngine is a script engine that will load script file and manage images, texts, effects and so on

Widget

wMiniScriptEngine :: Given StyleSheet => WConfig "script_engine" -> GameM (Widget Op'MiniScriptEngine)

Method

type Op'MiniScriptEngine = '[Op'Reset (), Op'Render, Op'Run, Op'HandleEvent, Op'Switch, Op'LoadMiniScript, Op'Goto, Op'SetFields]

Operators

op'loadMiniScript :: (∈) Op'LoadMiniScript xs => MiniScript () -> Getter (Widget xs) (Widget xs)

data Op'Goto br m r

op'goto :: (∈) Op'Goto xs => String -> Getter (Widget xs) (GameM (Widget xs))

data Op'SetFields br m r

op'setFields :: (∈) Op'SetFields xs => [(String, String)] -> Getter (Widget xs) (Widget xs)

Script

parseMiniScript :: FilePath -> IO (Maybe (MiniScript ()))

isReturn :: Skeleton t a -> Bool

(=:) :: a -> b -> (a, b) infix 2

data EffectIn

Constructors

FadeInLinear Int

FadeInEase Int

CharacterIn Int

NoEffectIn

data EffectOut

Constructors

FadeOutLinear Int

FadeOutEase Int

CharacterOut Int

NoEffectOut

type MiniScript = Skeleton SimpleDSL

wait :: Int -> MiniScript ()

destroyImage :: RefImage -> MiniScript ()

withEffect :: EffectIn -> EffectOut -> RefImage -> MiniScript a -> MiniScript a

data Position

Constructors

L

C

R

loadCharacter :: FilePath -> Position -> MiniScript RefImage

showCharacter :: RefImage -> MiniScript () -> MiniScript ()

speak :: Maybe RefImage -> [String] -> MiniScript ()

resetOpacity :: MiniScript ()

Orphan instances

Conf Symbol "script_engine"

Associated Types

type Required "script_engine" (k :: "script_engine") :: [Assoc Symbol *]

type Optional "script_engine" (k :: "script_engine") :: [Assoc Symbol *]

Methods

def :: Record Symbol (Optional "script_engine" k)