SDLight.Widgets.TabSelector

Description

Widget with tabs and each tab contains a SDLight.Widgets.Selector

Widget

wTabSelector :: Given StyleSheet => WConfig "tab_selector" -> Widget Op'TabSelector

TabSelector widget

Config Parameter

Required

[]

Optional

[ "selectNum" >: Int  -- How many items can I select?
, "pager" >: Maybe Int  -- Number of lines that single page can contain
]

Methods

op'reset Reset operator

op'render Render operator; dropdown style

op'renderTabSelector Render operator with custom renderer

op'run Run operator

op'handleEvent Keyevent handle operator

op'switch Check if this widget is alive/dead

op'getSelecting Get selecting indices as integer list

op'getPointer Get the current pointer index

op'getCurrentSelector Get the current selector (named-)widget, this returns Nothing when current pointer is not initialized

op'getTabName Get name of the current tab

op'setTabs Set tabs by a list of a pair of tabname and selector labels

wTabSelectLayer :: Given StyleSheet => WConfig "tab_select_layer" -> GameM (Widget Op'TabSelectLayer)

TabSelector widget with window layer and cursor layer

Config Parameter

Required

[ "windowTexture" >: SDL.Texture  -- Texture of window layer
, "cursorTexture" >: SDL.Texture  -- Texture of cursor layer
, "size" >: V2 Int  -- Size of this widget
]

Optional

[ "tabWidth" >: Int  -- Width of each tab
, "selectNum" >: Int  -- How many items can I select?
, "pager" >: Maybe Int  -- Number of lines that single page can contain
]

Methods

op'reset Reset operator

op'render Render operator; dropdown style

op'renderTabSelector Render operator with custom renderer

op'run Run operator

op'handleEvent Keyevent handle operator

op'switch Check if this widget is alive/dead

op'getSelecting Get selecting indices as integer list

op'getPointer Get the current pointer index

op'getCurrentSelector Get the current selector (named-)widget, this returns Nothing when current pointer is not initialized

op'getTabName Get name of the current tab

op'setTabs Set tabs by a list of a pair of tabname and selector labels

Method

type Op'TabSelector = '[Op'Reset (), Op'Render, Op'RenderTabSelector, Op'Run, Op'HandleEvent, Op'Switch, Op'GetSelecting, Op'GetPointer, Op'GetCurrentSelector, Op'GetTabName, Op'SetTabs]

Methods of wTabSelector

type Op'TabSelectLayer = '[Op'Reset (), Op'Render, Op'Run, Op'HandleEvent, Op'Switch, Op'GetSelecting, Op'GetPointer, Op'GetCurrentSelector, Op'GetTabName, Op'SetTabs]

Methods of wTabSelectLayer

Operator

op'getTabName :: (∈) Op'GetTabName xs => Getter (Widget xs) (Maybe String)

op'renderTabSelector :: (∈) Op'RenderTabSelector xs => (TabSelectorRenderConfig -> SelectorRenderConfig -> GameM ()) -> Getter (Widget xs) (GameM ())

op'getCurrentSelector :: (∈) Op'GetCurrentSelector xs => Getter (Widget xs) (Maybe (NamedWidget Op'Selector))

op'setTabs :: (∈) Op'SetTabs xs => [(String, [String])] -> Getter (Widget xs) (Widget xs)

Config Parameter

type TabSelectorRenderConfig = Record '["tabName" >: String, "tabIndex" >: Int, "isTabSelected" >: Bool]

Custom Renderer

Orphan instances

Conf Symbol "tab_select_layer"

Associated Types

type Required "tab_select_layer" (k :: "tab_select_layer") :: [Assoc Symbol *]

type Optional "tab_select_layer" (k :: "tab_select_layer") :: [Assoc Symbol *]

Methods

def :: Record Symbol (Optional "tab_select_layer" k)

Conf Symbol "tab_selector"

Associated Types

type Required "tab_selector" (k :: "tab_selector") :: [Assoc Symbol *]

type Optional "tab_selector" (k :: "tab_selector") :: [Assoc Symbol *]

Methods

def :: Record Symbol (Optional "tab_selector" k)